+++ R/GitHostGitLab.R
@@ -95,31 +95,12 @@ GitHostGitLab <- R6::R6Class("GitHostGitLab",
     # Set groups endpoint
     set_orgs_endpoint = function() {
       private$endpoints$orgs <- glue::glue("{private$api_url}/groups")
+      private$endpoints$users <- glue::glue("{private$api_url}/users?username=")
     },
 
     # Set projects endpoint
     set_repositories_endpoint = function() {
       private$endpoints$repositories <- glue::glue("{private$api_url}/projects")
-    },
-
-    # Set owner type
-    set_owner_type = function(owners) {
-      graphql_engine <- private$engines$graphql
-      user_or_org_query <- graphql_engine$gql_query$user_or_org_query
-      login_types <- purrr::map(owners, function(owner) {
-        response <- graphql_engine$gql_response(
-          gql_query = user_or_org_query,
-          vars = list(
-            "username"  = owner,
-            "grouppath" = owner
-          )
-        )
-        type <- purrr::discard(response$data, is.null) %>%
-          names()
-        attr(owner, "type") <- type
-        return(owner)
-      })
-      return(login_types)
     },
 
     # Setup REST and GraphQL engines
@@ -224,7 +205,8 @@ GitHostGitLab <- R6::R6Class("GitHostGitLab",
                                       progress = verbose) {
       if ("repo" %in% private$searching_scope) {
         rest_engine <- private$engines$rest
-        orgs <- private$set_owner_type(
+        graphql_engine <- private$engines$graphql
+        orgs <- graphql_engine$set_owner_type(
           owners = names(private$orgs_repos)
         )
         commits_table <- purrr::map(orgs, function(org) {
@@ -290,7 +272,7 @@ GitHostGitLab <- R6::R6Class("GitHostGitLab",
                                             progress = TRUE) {
       if ("repo" %in% private$searching_scope) {
         graphql_engine <- private$engines$graphql
-        orgs <- private$set_owner_type(
+        orgs <- graphql_engine$set_owner_type(
           owners = names(private$orgs_repos)
         )
         files_table <- purrr::map(orgs, function(org) {