+++ tests/testthat/test-01-get_repos-GitHub.R
@@ -568,7 +568,7 @@ test_that("get_repos_from_repos works", {
   attr(test_org, "type") <- "organization"
   mockery::stub(
     github_testhost_priv$get_repos_from_repos,
-    "private$set_owner_type",
+    "graphql_engine$set_owner_type",
     test_org
   )
   mockery::stub(